Commit five
This commit is around 10hours late or so. In this commit I fixed some
issues with flex not working properly on the index.html, contact.html
pages. I also changes the layout of the contact.html page so the form is
now in the center of the page to make it feel less empty. I added more
of a proper heading and some short text to the form as well.