Commit Six
In this commit I worked on the cv.html page and the bio.html page, I
gave them both a layout and styled them somewhat. More work needs to be
done in terms of styling, especially for the Bio.html page, things such
as icons could also be added to help improve user engagement and make
the pages more appealing to look at too.